[Refactor] BookRecommendationKit 내부 로직 SRP 원칙 준수토록 파일 분리 (#66)
- 최대 반환토큰값 초과로 GPTFormat 에러 반환하는 에러 수정
