+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Services/ImageDownload/ImageDownloadAPI.swift
@@ -1,5 +1,4 @@
 import BookMatchCore
-import Foundation
 import NetworkKit
 import RxSwift
 import UIKit
