+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Services/OpenAI/Endpoint/OpenAIEndpoint.swift
@@ -4,8 +4,6 @@ import NetworkKit
 struct ChatGPTEndpoint: Endpoint {
     // MARK: - Nested Types
 
-    // MARK: - Internal
-
     typealias Response = ChatGPTResponse
 
     // MARK: - Properties
