+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchCore/BookMatchError/BookMatchError.swift
@@ -2,7 +2,7 @@
 public enum BookMatchError: Error {
     case noMatchFound
     case networkError
-    case invalidGPTFormat
+    case invalidGPTFormat(String)
     case OCRError(String)
     case CoreMLError(String)
     case imageCalculationFailed(String)
@@ -20,8 +20,8 @@ public enum BookMatchError: Error {
             return "검색하신 조건에 맞는 책을 찾지 못했습니다"
         case .networkError:
             return "네트워크 연결이 원활하지 않습니다\n잠시 후 다시 시도해주세요"
-        case .invalidGPTFormat:
-            return "일시적인 오류가 발생했습니다\n잠시 후 다시 시도해주세요"
+        case let .invalidGPTFormat(response):
+            return "GPT 반환 format 에러 / 결과: \(response)"
         case .imageCalculationFailed:
             return "이미지 유사도 연산에 실패했습니다.\n잠시 후 다시 시도해주세요"
         case .imageDownloadFailed:
